pyproject.toml
[project]
name = "andgein2026"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.5",
    "google-auth>=2.0.0",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "transformers[torch]>=4.57.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "opencv-python>=4.12.0.88",
    "ftfy>=6.3.1",
    "scipy>=1.16.3",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
]
